refactor: Refactorización por principio de rensponsabilidad unica
- Se aplicó el Principio de Responsabilidad Única en los metodos que realizaban multiples rensponsabilidades

+++ src/gestortutoriasfx/controlador/FXMLFormularioEvidenciaController.java
@@ -163,7 +163,7 @@ public class FXMLFormularioEvidenciaController implements Initializable {
             return false;
         }
 
-        long limiteBytes = 5 * 1024 * 1024;
+        long limiteBytes = ((long) 5 * 1024 * 1024);
         
         if (archivo.length() > limiteBytes) {
             Utilidades.mostrarAlertaSimple("Archivo pesado",
